:adhesive_bandage: [#1771] -- Checking if form definition isn't alerady deleted

+++ src/openforms/forms/models/form_step.py
@@ -82,7 +82,7 @@ class FormStep(OrderedModel):
     def delete(self, using=None, keep_parents=False):
         super().delete(using=using, keep_parents=keep_parents)
 
-        if not self.form_definition.is_reusable:
+        if self.form_definition.pk is not None and not self.form_definition.is_reusable:
             self.form_definition.delete()
 
     def iter_components(self, recursive=True, **kwargs):
